+++ curve_editors/uv/vf_curve_manual_editor.py
@@ -25,8 +25,8 @@ class ManualCurveEdit:
 
 def editable_anchor_from_profile(profile: dict) -> tuple[int, int] | None:
     try:
-        voltage_mv = int(round(float(profile.get("candidate_voltage_mv"))))
-        clock_mhz = int(round(float(profile.get("lock_clock_mhz"))))
+        voltage_mv = round(float(profile.get("candidate_voltage_mv")))
+        clock_mhz = round(float(profile.get("lock_clock_mhz")))
     except (TypeError, ValueError):
         return None
     if voltage_mv <= 0 or clock_mhz <= 0:
@@ -50,7 +50,7 @@ def manual_drag_anchor_edit(
 ) -> ManualCurveEdit:
     points = _sorted_plan(plan)
     anchor_index = _nearest_voltage_index(points, int(anchor_voltage_mv))
-    requested_index = _nearest_voltage_index(points, int(round(float(requested_voltage_mv))))
+    requested_index = _nearest_voltage_index(points, round(float(requested_voltage_mv)))
     max_index = (
         len(points) - 1
         if max_voltage_bins is None
@@ -121,7 +121,7 @@ def manual_flatten_from_existing_point(
     anchor_index = _nearest_voltage_index(points, int(anchor_voltage_mv))
     clicked_index = _nearest_voltage_index(
         points,
-        int(round(float(clicked_voltage_mv))),
+        round(float(clicked_voltage_mv)),
     )
     flatten_index = min(int(anchor_index), int(clicked_index))
     flatten_clock_mhz = int(points[flatten_index]["target_mhz"])
@@ -155,7 +155,7 @@ def manual_add_curve_point_edit(
     )
     target_index = _nearest_voltage_index(
         points,
-        int(round(float(requested_voltage_mv))),
+        round(float(requested_voltage_mv)),
     )
     if target_index == anchor_index:
         return ManualCurveEdit(
@@ -228,7 +228,7 @@ def manual_tune_single_point_edit(
     anchor_index = _nearest_voltage_index(points, int(anchor_voltage_mv))
     source_index = _nearest_voltage_index(
         points,
-        int(round(float(point_voltage_mv))),
+        round(float(point_voltage_mv)),
     )
     if source_index < anchor_index:
         region_first_index = 0
@@ -259,7 +259,7 @@ def manual_tune_single_point_edit(
     target_index = _clamp(
         _nearest_voltage_index(
             points,
-            int(round(float(requested_point_voltage_mv))),
+            round(float(requested_point_voltage_mv)),
         ),
         region_first_index,
         region_last_index,
@@ -436,7 +436,7 @@ def manual_select_curve_point(
     anchor_index = _nearest_voltage_index(points, int(edit.anchor_voltage_mv))
     selected_index = _nearest_voltage_index(
         points,
-        int(round(float(voltage_mv))),
+        round(float(voltage_mv)),
     )
     selected_voltage_mv = int(points[selected_index]["voltage_mv"])
     selected_clock_mhz = (
@@ -544,7 +544,7 @@ def manual_offset_selected_range(
         _nearest_voltage_index(points, int(range_start_voltage_mv)),
     )
     reference_index = _clamp(
-        _nearest_voltage_index(points, int(round(float(reference_voltage_mv)))),
+        _nearest_voltage_index(points, round(float(reference_voltage_mv))),
         int(start_index),
         int(anchor_index),
     )
@@ -891,7 +891,7 @@ def _normalized_control_voltage_mvs(
     seen: set[int] = set()
     for raw_voltage_mv in control_voltage_mvs:
         try:
-            voltage_mv = int(round(float(raw_voltage_mv)))
+            voltage_mv = round(float(raw_voltage_mv))
         except (TypeError, ValueError):
             continue
         index = _nearest_voltage_index(points, voltage_mv)
